+++ common/src/main/java/whocraft/tardis_refined/common/block/shell/GlobalShellBlock.java
@@ -2,13 +2,17 @@ package whocraft.tardis_refined.common.block.shell;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
+import net.minecraft.server.level.ServerChunkCache;
 import net.minecraft.server.level.ServerLevel;
+import net.minecraft.util.RandomSource;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.BlockEntity;
@@ -26,10 +30,14 @@ import org.jetbrains.annotations.Nullable;
 import whocraft.tardis_refined.TRConfig;
 import whocraft.tardis_refined.common.blockentity.shell.GlobalShellBlockEntity;
 import whocraft.tardis_refined.common.blockentity.shell.ShellBaseBlockEntity;
+import whocraft.tardis_refined.common.capability.tardis.TardisLevelOperator;
 import whocraft.tardis_refined.common.tardis.themes.ShellTheme;
 import whocraft.tardis_refined.compat.ModCompatChecker;
 import whocraft.tardis_refined.compat.portals.ImmersivePortals;
 import whocraft.tardis_refined.compat.valkyrienskies.VSHelper;
+import whocraft.tardis_refined.registry.TRUpgrades;
+
+import java.util.Optional;
 
 public class GlobalShellBlock extends ShellBaseBlock {
 
@@ -47,6 +55,46 @@ public class GlobalShellBlock extends ShellBaseBlock {
     protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
         super.createBlockStateDefinition(builder);
         builder.add(LIT);
+    }
+
+    private Optional<TardisLevelOperator> getTARDIS(Level level, BlockPos blockPos) {
+        if (!level.isClientSide() && level.getBlockEntity(blockPos) instanceof GlobalShellBlockEntity shell) {
+            var interior = level.getServer().getLevel(shell.getTardisId());
+            if (interior != null) {
+                return TardisLevelOperator.get(interior);
+            }
+        }
+        return Optional.empty();
+    }
+
+    @Override
+    public void onPlace(BlockState blockState, Level level, BlockPos blockPos, BlockState blockState2, boolean bl) {
+        super.onPlace(blockState, level, blockPos, blockState2, bl);
+        level.scheduleTick(blockPos, blockState.getBlock(), 1); // The shell won't know which TARDIS it's linked to yet, so we check landing on the next tick.
+    }
+
+    @Override
+    public void randomTick(BlockState state, ServerLevel level, BlockPos pos, RandomSource random) {
+
+    }
+
+    @Override
+    public void tick(BlockState state, ServerLevel level, BlockPos pos, RandomSource random) {
+        var tardis = getTARDIS(level, pos);
+        if (tardis.isEmpty()) return;
+        // We only check for the presence of the upgrade to allow players to enable it in the settings menu while they're landing.
+        if (tardis.get().getPilotingManager().isLanding() && tardis.get().getPilotingManager().isInFlight() && TRUpgrades.MATERIALIZE_AROUND.get().isUnlocked(tardis.get().getUpgradeHandler())) {
+            // TODO This fallback can probably be removed in 26.1+
+            if (!level.getChunkSource().chunkMap.anyPlayerCloseEnoughForSpawning(new ChunkPos(pos))) {
+                level.getEntitiesOfClass(
+                        Entity.class,
+                        state.getCollisionShape(level, pos).bounds().move(pos)
+                ).forEach(entity -> entityInside(state, level, pos, entity));
+            }
+            // Scheduled ticks are a useful and simple way to check for entities in unloaded chunks.
+            // This also won't enable mob spawning (unlike a forced ticket), which makes accidental landings around hostile mobs less likely.
+            level.scheduleTick(pos, state.getBlock(), 1);
+        }
     }
 
     @Override