+++ common/src/main/java/whocraft/tardis_refined/common/tardis/manager/TardisPilotingManager.java
@@ -57,7 +57,7 @@ public class TardisPilotingManager extends TickableHandler {
 
     public static final int MAX_THROTTLE_STAGE = 5;
     // CONSTANTS
-    private static final int TICKS_LANDING_MAX = 9 * 20;
+    public static final int TICKS_LANDING_MAX = 9 * 20;
     public static final int TICKS_COOLDOWN_MAX = (10 * 60) * 20;
     private static final double DEFAULT_MAXIMUM_FUEL = 1000;
     private static final double FLIGHT_COST = 0.5f;