+++ changelog.md
@@ -31,3 +31,4 @@
 - Fixed items dropped in shell view ending up outside the TARDIS.
 - Fixed players being able to pick up items and XP while in shell view.
 - Fixed the TARDIS sometimes playing the default hum when set to a different hum.
+- Fixed materialize around not working in unloaded chunks.